check message type first

+++ app/Http/Controllers/UsersController.php
@@ -141,14 +141,18 @@ class UsersController extends Controller
             $subscriber = Subscriber::where('telephone', '=', $to)->first();
             if ($subscriber) {
                 Log::debug(print_r($data, true));
-                if ($data->entry[0]->changes[0]->value->messages[0]->text->body === 'أوقف الخدمة') {
-                    $myservice = new MyService;
-                    $myservice->sendWhatsApp($subscriber, '', 'stop_service');
-                    $subscriber = Subscriber::where('telephone', '=', $to);
-                    $subscriber->delete();
+                if ($data->entry[0]->changes[0]->value->messages[0]->type == "text") {
+                    if ($data->entry[0]->changes[0]->value->messages[0]->text->body === 'أوقف الخدمة') {
+                        $myservice = new MyService;
+                        $myservice->sendWhatsApp($subscriber, '', 'stop_service');
+                        $subscriber = Subscriber::where('telephone', '=', $to);
+                        $subscriber->delete();
+                    } else {
+                        $myservice = new MyService;
+                        $myservice->sendWhatsApp($subscriber, '', 'to_stop_service');
+                    }
                 } else {
-                    $myservice = new MyService;
-                    $myservice->sendWhatsApp($subscriber, '', 'to_stop_service');
+                    // message is not text
                 }
             } else {
                 // user no longer a subscriber
